Go back to authors and docs split
it is not better or worse than the new one, but it breaks less code
